KSecret file serialization logic rewrite
Abandoned the parts format
A special device is introduced allowing custom reads/writes
All the encryption logic moved into a dedicated class
KSecretEncryptionFilter
All tests pass but information is stored unencrypted for the moment.
This is for debugging reasons. I'm now working on the encryption filter.

Also, the daemon jobs switched to KCompositeJob